backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
requests
yt-dlp
faster-whisper
pydantic
